# TestTaskGolang/docs/swagger.yaml
basePath: /
definitions:
  handler.createRequest:
    properties:
      end_date:
        example: 2025-12
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 2025-07
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для работы с подписками
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: читает подписки
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: прочитать подписки
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Добавляет новую подписку в систему
      parameters:
      - description: Subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handler.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Возвращает статус
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку по ID
      tags:
      - subscriptions
    get:
      description: Возвращает одну подписку
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID пользователя
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Изменяет данные подписки по ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/handler.createRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Возвращает суммарную стоимость всех подписок за период с фильтрацией
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - description: Service name
        in: query
        name: service_name
        required: true
        type: string
      - description: Start period (YYYY-MM)
        in: query
        name: from
        required: true
        type: string
      - description: End period (YYYY-MM)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Подсчитать сумму подписок
      tags:
      - subscriptions
swagger: "2.0"